requirements.txt
# requirements.txt
brainflow==5.10.0
pyserial==3.5
pyyaml==6.0.1
pylsl==1.16.1
numpy==1.24.3